+++ packages/rspack/tests/configCases/css/css-cross-origin-loading/index.js
@@ -2,5 +2,5 @@ it("should load css chunk without crossOrigin", async function () {
   import("./chunk.css").catch(() => { });
   expect(document.head._children[0]._type).toBe("link");
   expect(document.head._children[0]._href).toBe("https://test.cases/path/19.bundle0.css");
-  expect(document.head._children[0].crossOrigin).toBeUndefined();
+  expect(document.head._children[0].crossOrigin).toBeFalsy();
 });
