+++ packages/rspack/tests/configCases/nonce/not-set-nonce/index.js
@@ -6,7 +6,7 @@ it("should load script without nonce", function () {
 
 	var script = document.head._children.pop();
 	__non_webpack_require__("./chunk-without-nonce.web.js");
-	expect(script.getAttribute("nonce")).toBeUndefined();
+	expect(script.getAttribute("nonce")).toBeFalsy();
 
 	return promise;
 });
