+++ packages/rspack/tests/hotCases/chunk/multi-chunk-single-runtime/snapshot/web/1.snap.txt
@@ -1,0 +1,116 @@
+# Case multi-chunk-single-runtime: Step 1
+
+## Changed Files
+- file.js
+
+## Asset Files
+- Bundle: a.js
+- Bundle: b.js
+- Bundle: main.js
+- Bundle: main_async_js.chunk.CURRENT_HASH.js
+- Bundle: runtime.js
+- Manifest: runtime.LAST_HASH.hot-update.json, size: 46
+- Update: a.LAST_HASH.hot-update.js, size: 94
+- Update: b.LAST_HASH.hot-update.js, size: 94
+- Update: main.LAST_HASH.hot-update.js, size: 97
+- Update: runtime.LAST_HASH.hot-update.js, size: 204
+
+## Manifest
+
+### runtime.LAST_HASH.hot-update.json
+
+```json
+{"c":["a","b","main","runtime"],"r":[],"m":[]}
+```
+
+		
+## Update
+
+
+### a.LAST_HASH.hot-update.js
+
+#### Changed Modules
+- ./file.js
+
+#### Changed Runtime Modules
+
+
+#### Changed Content
+```js
+self["webpackHotUpdate"]('a', {
+"./file.js": (function (module) {
+module.exports = 2;
+}),
+
+});
+```
+
+
+
+
+### b.LAST_HASH.hot-update.js
+
+#### Changed Modules
+- ./file.js
+
+#### Changed Runtime Modules
+
+
+#### Changed Content
+```js
+self["webpackHotUpdate"]('b', {
+"./file.js": (function (module) {
+module.exports = 2;
+}),
+
+});
+```
+
+
+
+
+### main.LAST_HASH.hot-update.js
+
+#### Changed Modules
+- ./file.js
+
+#### Changed Runtime Modules
+
+
+#### Changed Content
+```js
+self["webpackHotUpdate"]('main', {
+"./file.js": (function (module) {
+module.exports = 2;
+}),
+
+});
+```
+
+
+
+
+### runtime.LAST_HASH.hot-update.js
+
+#### Changed Modules
+
+
+#### Changed Runtime Modules
+- webpack/runtime/get_full_hash
+
+#### Changed Content
+```js
+self["webpackHotUpdate"]('runtime', {
+
+},function(__webpack_require__) {
+// webpack/runtime/get_full_hash
+!function() {
+__webpack_require__.h = function () {
+	return "CURRENT_HASH";
+};
+
+}();
+
+}
+);
+```
