+++ packages/rspack/tests/hotCases/plugins/banner/test.filter.js
@@ -1,0 +1,1 @@
+module.exports = () => false;
